backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
django-cors-headers>=4.3
django-filter>=24.0
mysqlclient>=2.2
python-decouple>=3.8
Pillow>=10.0
reportlab>=4.0
